ファイル移動、__name__ == "__main__"を追加
